requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
Flask-Cors>=3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.6
gunicorn==21.2.0
newrelic>=8.10.1
psycopg2-binary==2.9.1
